requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.114.0
googleapis-common-protos==1.59.1
greenlet==3.0.3
grpcio==1.66.1
h11==0.14.0
idna==3.8
importlib_metadata==8.4.0
JSON-log-formatter==1.0
opentelemetry-api==1.27.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
protobuf==4.25.4
psycopg2==2.9.9
pydantic==2.9.0
pydantic_core==2.23.2
python-multipart==0.0.9
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.4
thrift==0.20.0
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
wrapt==1.16.0
zipp==3.20.1
